rubric-discovery/backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
websockets>=12.0
openai>=1.12
python-dotenv>=1.0
numpy>=1.24
